+++ packages/contracts/contracts/voting/token/TokenVotingSetup.sol
@@ -33,8 +33,8 @@ contract TokenVotingSetup is PluginSetup {
     /// @notice The address of the `TokenVoting` base contract.
     TokenVoting private immutable tokenVotingBase;
 
-    /// @notice The address zero to be used as oracle address for permissions.
-    address private constant NO_ORACLE = address(0);
+    /// @notice The address zero to be used as condition address for permissions.
+    address private constant NO_CONDITION = address(0);
 
     /// @notice The address of the `GovernanceERC20` base contract.
     address public immutable governanceERC20Base;
@@ -194,7 +194,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Grant,
             plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
@@ -202,7 +202,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Grant,
             plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
         );
 
@@ -211,7 +211,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Grant,
             _dao,
             plugin,
-            NO_ORACLE,
+            NO_CONDITION,
             DAO(payable(_dao)).EXECUTE_PERMISSION_ID()
         );
 
@@ -222,7 +222,7 @@ contract TokenVotingSetup is PluginSetup {
                 PermissionLib.Operation.Grant,
                 token,
                 _dao,
-                NO_ORACLE,
+                NO_CONDITION,
                 tokenMintPermission
             );
         }
@@ -265,7 +265,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Revoke,
             _plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
@@ -273,7 +273,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Revoke,
             _plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
         );
 
@@ -281,7 +281,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Revoke,
             _dao,
             _plugin,
-            NO_ORACLE,
+            NO_CONDITION,
             DAO(payable(_dao)).EXECUTE_PERMISSION_ID()
         );
 
@@ -295,7 +295,7 @@ contract TokenVotingSetup is PluginSetup {
                 PermissionLib.Operation.Revoke,
                 token,
                 _dao,
-                NO_ORACLE,
+                NO_CONDITION,
                 GovernanceERC20(token).MINT_PERMISSION_ID()
             );
         }