+++ packages/contracts/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed `_canVote` logic to revert for votes being cast with `VoteOption.None`.
 - Changed `callId` in `DAO.execute()` from `uint256` to `bytes32`.
 - Changed the `createProposal` function signature which expects `uint256` `allowFaiureMap` argument.
+- Changed all occurences of `oracle` to `condition`.
 
 ### Removed